requirements-txt.txt
# Core dependencies
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0

# API and async
aiohttp>=3.9.0
requests>=2.31.0
backoff>=2.2.0

# Machine Learning
scikit-learn>=1.3.0
xgboost>=2.0.0
joblib>=1.3.0
scipy>=1.11.0

# Data visualization
plotly>=5.17.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Database
sqlalchemy>=2.0.0
aiosqlite>=0.19.0

# Utilities
pydantic>=2.4.0
typing-extensions>=4.8.0
python-dateutil>=2.8.0

# Logging and monitoring
colorlog>=6.7.0
python-json-logger>=2.0.0

# Testing (optional, for development)
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Code quality (optional, for development)
black>=23.9.0
flake8>=6.1.0
mypy>=1.5.0
isort>=5.12.0

# Documentation (optional)
sphinx>=7.2.0
sphinx-rtd-theme>=1.3.0